requirements.txt
# ===============================================
# PYTHON DEPENDENCIES
# CASML RAG System Requirements
# ===============================================

# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1

# NLP & Text Processing
nltk==3.8.1
spacy==3.6.1
langchain==0.1.0
langchain-community==0.0.10

# Embeddings & Transformers (TensorFlow)
tensorflow==2.15.0
# tensorflow-gpu==2.15.0  # Use this if you have CUDA GPU
transformers==4.35.0
sentence-transformers==2.2.2
tensorflow-hub==0.15.0
tensorflow-text==2.15.0

# Vector Stores & Indexing
faiss-cpu==1.7.4  # TensorFlow compatible
# faiss-gpu==1.7.4  # Use if you have CUDA GPU
chromadb==0.4.18
rank-bm25==0.2.2
scann==1.3.1  # Google's vector similarity search (TensorFlow-optimized)

# Retrieval & Re-ranking
beir==2.0.0  # For retrieval evaluation
pymilvus==2.3.3  # Alternative vector DB

# LLM Integration
openai==1.3.0  # If using OpenAI API
anthropic==0.7.0  # If using Claude API
huggingface-hub==0.19.4

# Evaluation Metrics
rouge-score==0.1.2
sacrebleu==2.3.1
bert-score==0.3.13
evaluate==0.4.1

# Data Processing
pypdf==3.17.0  # PDF parsing
python-docx==1.1.0  # Word document parsing
beautifulsoup4==4.12.2  # HTML parsing
lxml==4.9.3

# Experiment Tracking & Logging
wandb==0.16.0
mlflow==2.9.0
tensorboard==2.15.0

# Utilities
scikit-learn==1.3.2
jsonlines==4.0.0
datasets==2.14.6  # HuggingFace datasets
kaggle==1.5.16  # Kaggle API

# Testing & Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0
jupyter==1.0.0
ipykernel==6.26.0
